requirements.txt
streamlit
langchain
langchain-text-splitters
langchain-google-genai
langchain-community
faiss-cpu
supadata
python-dotenv